[Feat] 마이페이지 내가 만들었던 방 히스토리 보기 + 타인 방 히스토리

See merge request s13-webmobile1-sub1/S13P11A404!244

+++ backend/BE/src/main/java/com/a404/duckonback/service/RoomService.java
@@ -7,10 +7,10 @@ import java.util.Optional;
 
 public interface RoomService {
     Room createRoom(Room room);
-    Optional<Room> getRoomById(Integer roomId);
+    Optional<Room> getRoomById(Long roomId);
     List<Room> getAllRooms();
-    Room updateRoom(Integer roomId, Room updatedRoom);
-    void deleteRoom(Integer roomId);
+    Room updateRoom(Long roomId, Room updatedRoom);
+    void deleteRoom(Long roomId);
     List<Room> getRoomsByCreator(Long id);
-    List<Room> getRoomsByArtist(Integer artistId);
+    List<Room> getRoomsByArtist(Long artistId);
 }
